# swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Boltz API",
    "version": "3.12.1"
  },
  "paths": {
...
    "/funding": {
      "post": {
        "description": "Create a new funding address for pre-funding swaps",
        "tags": [
          "Funding Address"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FundingAddressRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created funding address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundingAddressResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error that caused the funding address creation to fail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "When the currency is not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/funding/{id}/signature": {
      "get": {
        "tags": [
          "Funding Address"
        ],
        "description": "Get the signing details for a funding address to create a swap",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the funding address"
          },
          {
            "in": "query",
            "name": "swapId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the swap to fund with this funding address"
          }
        ],
        "responses": {
          "200": {
            "description": "Signing details for the funding transaction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundingAddressSigningDetails"
                }
              }
            }
          },
          "400": {
            "description": "Error that caused the request to fail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "When the funding address is not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Funding Address"
        ],
        "description": "Submit a partial signature for a funding address transaction",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the funding address"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartialSignature"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Signature was accepted and transaction broadcast"
          },
          "400": {
            "description": "Error that caused the signature submission to fail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "When the funding address is not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PartialSignature": {
        "type": "object",
        "required": [
          "pubNonce",
          "partialSignature"
        ],
        "properties": {
          "pubNonce": {
            "type": "string",
            "description": "Public nonce encoded as HEX"
          },
          "partialSignature": {
            "type": "string",
            "description": "Partial signature encoded as HEX"
          }
        }
      },
...
      "FundingAddressRequest": {
        "type": "object",
        "required": [
          "symbol",
          "refundPublicKey"
        ],
        "properties": {
          "symbol": {
            "type": "string",
            "description": "Currency symbol for the funding address (e.g. BTC, L-BTC)"
          },
          "refundPublicKey": {
            "type": "string",
            "description": "Public key with which the funding address can be refunded encoded as HEX"
          }
        }
      },
...
      "FundingAddressResponse": {
        "type": "object",
        "required": [
          "id",
          "address",
          "timeoutBlockHeight",
          "boltzPublicKey"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "ID of the created funding address"
          },
          "address": {
            "type": "string",
            "description": "Address to which funds should be sent"
          },
          "timeoutBlockHeight": {
            "type": "number",
            "description": "Block height at which the funding address times out"
          },
          "boltzPublicKey": {
            "type": "string",
            "description": "Public key of Boltz used in the funding address encoded as HEX"
          }
        }
      },
      "FundingAddressSigningDetails": {
        "type": "object",
        "required": [
          "pubNonce",
          "publicKey",
          "transactionHex",
          "transactionHash"
        ],
        "properties": {
          "pubNonce": {
            "type": "string",
            "description": "Public nonce of Boltz for the signing session encoded as HEX"
          },
          "publicKey": {
            "type": "string",
            "description": "Public key of Boltz encoded as HEX"
          },
          "transactionHex": {
            "type": "string",
            "description": "Transaction to be signed encoded as HEX"
          },
          "transactionHash": {
            "type": "string",
            "description": "Hash of the transaction to be signed encoded as HEX"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Description of the error that caused the request to fail"
          }
        }
      }
    }
  },
...
  "servers": [
    {
      "url": "https://api.boltz.exchange/v2",
      "description": "Mainnet"
    },
    {
      "url": "http://localhost:9006/v2",
      "description": "Regtest"
    }
  ]
}